Fixed unit tests

+++ src/main/java/com/musala/gateway/dao/PeripheralDeviceDao.java
@@ -1,5 +1,5 @@
 package com.musala.gateway.dao;
 
 public interface PeripheralDeviceDao extends BaseDao {
-    void remove(int id);
+    void remove(long id);
 }
